vendor_platform/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==10.0.1
django-redis==5.2.0
redis==5.0.1
PyJWT==2.8.0
django-filter==23.3
django-storages==1.14.2
boto3==1.28.62
celery==5.3.4
django-celery-results==2.5.1
python-decouple==3.8
drf-yasg==1.21.5
Faker==19.6.2
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.2.0